requirements.txt
playwright
pdfplumber
openpyxl
pandas
requests
beautifulsoup4
python-dotenv
supabase
tabula-py
lxml
tenacity
